Moved all data types to separate folders in lib
