+++ lib/sweet-data.js
@@ -34,200 +34,30 @@ pp.prototype.setStep = function(steps, stepChar){
 	
 // ----------------------- XML section ----------------------------------------------------
 
-pp.prototype.xml = require('./xml/index').xml;
+pp.prototype.xml = require('./xml').xml;
 
 // ----------------------- JSON section ----------------------------------------------------
 
-pp.prototype.json = function(text) {
-
-	if ( typeof text === "string" ) {
-		return JSON.stringify(JSON.parse(text), null, this.step);
-	}
-	if ( typeof text === "object" ) {
-		return JSON.stringify(text, null, this.step);
-	}
-	return null;
-}
+pp.prototype.json = require('./json/').json;
 
 // ----------------------- CSS section ----------------------------------------------------
 
-pp.prototype.css = function(text) {
-
-	var ar = text.replace(/\s{1,}/g,' ')
-				.replace(/\{/g,"{~::~")
-				.replace(/\}/g,"~::~}~::~")
-				.replace(/\;/g,";~::~")
-				.replace(/\/\*/g,"~::~/*")
-				.replace(/\*\//g,"*/~::~")
-				.replace(/~::~\s{0,}~::~/g,"~::~")
-				.split('~::~'),
-		len = ar.length,
-		deep = 0,
-		str = '',
-		ix = 0;
-		
-		for(ix=0;ix<len;ix++) {
-
-			if( /\{/.exec(ar[ix]))  { 
-				str += this.shift[deep++]+ar[ix];
-			} else 
-			if( /\}/.exec(ar[ix]))  { 
-				str += this.shift[--deep]+ar[ix];
-			} else
-			if( /\*\\/.exec(ar[ix]))  { 
-				str += this.shift[deep]+ar[ix];
-			}
-			else {
-				str += this.shift[deep]+ar[ix];
-			}
-		}
-		return str.replace(/^\n{1,}/,'');
-}
+pp.prototype.css = require('./css').css;
 
 // ----------------------- SQL section ----------------------------------------------------
 
-function isSubquery(str, parenthesisLevel) {
-  return  parenthesisLevel - (str.replace(/\(/g,'').length - str.replace(/\)/g,'').length )
-}
 
-function split_sql(str, tab) {
 
-    return str.replace(/\s{1,}/g," ")
-
-        .replace(/ AND /ig,"~::~"+tab+tab+"AND ")
-        .replace(/ BETWEEN /ig,"~::~"+tab+"BETWEEN ")
-        .replace(/ CASE /ig,"~::~"+tab+"CASE ")
-        .replace(/ ELSE /ig,"~::~"+tab+"ELSE ")
-        .replace(/ END /ig,"~::~"+tab+"END ")
-        .replace(/ FROM /ig,"~::~FROM ")
-        .replace(/ GROUP\s{1,}BY/ig,"~::~GROUP BY ")
-        .replace(/ HAVING /ig,"~::~HAVING ")
-        //.replace(/ IN /ig,"~::~"+tab+"IN ")
-        .replace(/ IN /ig," IN ")
-        .replace(/ JOIN /ig,"~::~JOIN ")
-        .replace(/ CROSS~::~{1,}JOIN /ig,"~::~CROSS JOIN ")
-        .replace(/ INNER~::~{1,}JOIN /ig,"~::~INNER JOIN ")
-        .replace(/ LEFT~::~{1,}JOIN /ig,"~::~LEFT JOIN ")
-        .replace(/ RIGHT~::~{1,}JOIN /ig,"~::~RIGHT JOIN ")
-        .replace(/ ON /ig,"~::~"+tab+"ON ")
-        .replace(/ OR /ig,"~::~"+tab+tab+"OR ")
-        .replace(/ ORDER\s{1,}BY/ig,"~::~ORDER BY ")
-        .replace(/ OVER /ig,"~::~"+tab+"OVER ")
-        .replace(/\(\s{0,}SELECT /ig,"~::~(SELECT ")
-        .replace(/\)\s{0,}SELECT /ig,")~::~SELECT ")
-        .replace(/ THEN /ig," THEN~::~"+tab+"")
-        .replace(/ UNION /ig,"~::~UNION~::~")
-        .replace(/ USING /ig,"~::~USING ")
-        .replace(/ WHEN /ig,"~::~"+tab+"WHEN ")
-        .replace(/ WHERE /ig,"~::~WHERE ")
-        .replace(/ WITH /ig,"~::~WITH ")
-        //.replace(/\,\s{0,}\(/ig,",~::~( ")
-        //.replace(/\,/ig,",~::~"+tab+tab+"")
-        .replace(/ ALL /ig," ALL ")
-        .replace(/ AS /ig," AS ")
-        .replace(/ ASC /ig," ASC ") 
-        .replace(/ DESC /ig," DESC ") 
-        .replace(/ DISTINCT /ig," DISTINCT ")
-        .replace(/ EXISTS /ig," EXISTS ")
-        .replace(/ NOT /ig," NOT ")
-        .replace(/ NULL /ig," NULL ")
-        .replace(/ LIKE /ig," LIKE ")
-        .replace(/\s{0,}SELECT /ig,"SELECT ")
-        .replace(/~::~{1,}/g,"~::~")
-        .split('~::~');
-}
-
-pp.prototype.sql = function(text) {
-
-    var ar_by_quote = text.replace(/\s{1,}/g," ")
-                        .replace(/\'/ig,"~::~\'")
-                        .split('~::~'),
-        len = ar_by_quote.length,
-        ar = [],
-        deep = 0,
-        tab = this.step,//+this.step,
-        inComment = true,
-        inQuote = false,
-        parenthesisLevel = 0,
-        str = '',
-        ix = 0;
-
-    for(ix=0;ix<len;ix++) {
-
-        if(ix%2) {
-            ar = ar.concat(ar_by_quote[ix]);
-        } else {
-            ar = ar.concat(split_sql(ar_by_quote[ix], tab) );
-        }
-    }
-
-    len = ar.length;
-    for(ix=0;ix<len;ix++) {
-
-        parenthesisLevel = isSubquery(ar[ix], parenthesisLevel);
-
-        if( /\s{0,}\s{0,}SELECT\s{0,}/.exec(ar[ix]))  { 
-            ar[ix] = ar[ix].replace(/\,/g,",\n"+tab+tab+"")
-        } 
-
-        if( /\s{0,}\(\s{0,}SELECT\s{0,}/.exec(ar[ix]))  { 
-            deep++;
-            str += this.shift[deep]+ar[ix];
-        } else 
-        if( /\'/.exec(ar[ix]) )  { 
-            if(parenthesisLevel<1 && deep) {
-                deep--;
-            }
-            str += ar[ix];
-        }
-        else  { 
-            str += this.shift[deep]+ar[ix];
-            if(parenthesisLevel<1 && deep) {
-                deep--;
-            }
-        } 
-    }
-
-    str = str.replace(/^\n{1,}/,'').replace(/\n{1,}/g,"\n");
-    return str;
-}
+pp.prototype.sql = require('./sql').sql;
 
 // ----------------------- min section ----------------------------------------------------
 
-pp.prototype.xmlmin = require('./xml/index').xmlmin;
+pp.prototype.xmlmin = require('./xml').xmlmin;
 
-pp.prototype.jsonmin = function(text) {
-								  
-    return  text.replace(/\s{0,}\{\s{0,}/g,"{")
-                .replace(/\s{0,}\[$/g,"[")
-                .replace(/\[\s{0,}/g,"[")
-                .replace(/:\s{0,}\[/g,':[')
-                .replace(/\s{0,}\}\s{0,}/g,"}")
-                .replace(/\s{0,}\]\s{0,}/g,"]")
-                .replace(/\"\s{0,}\,/g,'",')
-                .replace(/\,\s{0,}\"/g,',"')
-                .replace(/\"\s{0,}:/g,'":')
-                .replace(/:\s{0,}\"/g,':"')
-                .replace(/:\s{0,}\[/g,':[')
-                .replace(/\,\s{0,}\[/g,',[')
-                .replace(/\,\s{2,}/g,', ')
-                .replace(/\]\s{0,},\s{0,}\[/g,'],[');   
-}
+pp.prototype.jsonmin = require('./json').jsonmin;
 
-pp.prototype.cssmin = function(text, preserveComments) {
-	
-	var str = preserveComments ? text
-				   : text.replace(/\/\*([^*]|[\r\n]|(\*+([^*/]|[\r\n])))*\*+\//g,"") ;
-	return str.replace(/\s{1,}/g,' ')
-			  .replace(/\{\s{1,}/g,"{")
-			  .replace(/\}\s{1,}/g,"}")
-			  .replace(/\;\s{1,}/g,";")
-			  .replace(/\/\*\s{1,}/g,"/*")
-			  .replace(/\*\/\s{1,}/g,"*/");
-}	
+pp.prototype.cssmin = require('./css').css;
 
-pp.prototype.sqlmin = function(text) {
-    return text.replace(/\s{1,}/g," ").replace(/\s{1,}\(/,"(").replace(/\s{1,}\)/,")");
-}
+pp.prototype.sqlmin = require('./sql').sqlmin;
 
 exports.SweetData = new pp;
